+++ programs/drift/src/math/amm_jit.rs
@@ -4,22 +4,21 @@ use crate::math::casting::Cast;
 use crate::math::constants::AMM_RESERVE_PRECISION;
 use crate::math::orders::standardize_base_asset_amount;
 use crate::math::safe_math::SafeMath;
-
-use crate::state::perp_market::PerpMarket;
-
+use crate::state::perp_market::{AMMLiquiditySplit, PerpMarket};
 #[cfg(test)]
 mod tests;
 
 // assumption: market.amm.amm_jit_is_active() == true
-// assumption: taker_baa will improve market balance (see orders.rs & amm_wants_to_make)
+// assumption: taker_baa will improve market balance (see orders.rs & amm_wants_to_jit_make)
 pub fn calculate_jit_base_asset_amount(
     market: &PerpMarket,
     maker_base_asset_amount: u64,
     auction_price: u64,
     valid_oracle_price: Option<i64>,
     taker_direction: PositionDirection,
+    liquidity_split: AMMLiquiditySplit,
 ) -> DriftResult<u64> {
-    // only take up to 50% of what the maker is making
+    // AMM can only take up to 50% of size the maker is offering
     let mut max_jit_amount = maker_base_asset_amount.safe_div(2)?;
 
     // check for wash trade
@@ -76,7 +75,8 @@ pub fn calculate_jit_base_asset_amount(
         return Ok(0);
     }
 
-    jit_base_asset_amount = calculate_clamped_jit_base_asset_amount(market, jit_base_asset_amount)?;
+    jit_base_asset_amount =
+        calculate_clamped_jit_base_asset_amount(market, liquidity_split, jit_base_asset_amount)?;
 
     jit_base_asset_amount = jit_base_asset_amount.min(max_jit_amount);
 
@@ -87,27 +87,98 @@ pub fn calculate_jit_base_asset_amount(
     Ok(jit_base_asset_amount)
 }
 
-// assumption: taker_baa will improve market balance (see orders.rs & amm_wants_to_make)
+// assumption: taker_baa will improve market balance (see orders.rs & amm_wants_to_jit_make)
 // note: we split it into two (calc and clamp) bc its easier to maintain tests
 pub fn calculate_clamped_jit_base_asset_amount(
     market: &PerpMarket,
+    liquidity_split: AMMLiquiditySplit,
     jit_base_asset_amount: u64,
 ) -> DriftResult<u64> {
     // apply intensity
     // todo more efficient method do here
-    let jit_base_asset_amount = jit_base_asset_amount
+    let jit_base_asset_amount: u64 = jit_base_asset_amount
         .cast::<u128>()?
-        .safe_mul(market.amm.amm_jit_intensity as u128)?
-        .safe_div(100)?
+        .safe_mul(market.amm.amm_jit_intensity.min(100).cast::<u128>()?)?
+        .safe_div(100_u128)?
         .cast::<u64>()?;
 
     // bound it; dont flip the net_baa
-    let max_amm_base_asset_amount = market
-        .amm
-        .base_asset_amount_with_amm
-        .unsigned_abs()
-        .cast::<u64>()?;
+    let max_amm_base_asset_amount = if liquidity_split != AMMLiquiditySplit::LPOwned {
+        market
+            .amm
+            .base_asset_amount_with_amm
+            .unsigned_abs()
+            .cast::<u64>()?
+    } else {
+        market
+            .amm
+            .imbalanced_base_asset_amount_with_lp()?
+            .unsigned_abs()
+            .cast::<u64>()?
+    };
+
     let jit_base_asset_amount = jit_base_asset_amount.min(max_amm_base_asset_amount);
 
     Ok(jit_base_asset_amount)
+}
+
+pub fn calculate_amm_jit_liquidity(
+    market: &mut PerpMarket,
+    taker_direction: PositionDirection,
+    maker_price: u64,
+    valid_oracle_price: Option<i64>,
+    base_asset_amount: u64,
+    taker_base_asset_amount: u64,
+    maker_base_asset_amount: u64,
+    taker_has_limit_price: bool,
+) -> DriftResult<(u64, AMMLiquiditySplit)> {
+    let mut jit_base_asset_amount: u64 = 0;
+    let mut liquidity_split: AMMLiquiditySplit = AMMLiquiditySplit::ProtocolOwned;
+
+    // taker has_limit_price = false means (limit price = 0 AND auction is complete) so
+    // market order will always land and fill on amm next round
+    let amm_will_fill_next_round: bool =
+        !taker_has_limit_price && maker_base_asset_amount < taker_base_asset_amount;
+
+    // return early
+    if amm_will_fill_next_round {
+        return Ok((jit_base_asset_amount, liquidity_split));
+    }
+    let amm_wants_to_jit_make = market.amm.amm_wants_to_jit_make(taker_direction)?;
+
+    let amm_lp_wants_to_jit_make = market.amm.amm_lp_wants_to_jit_make(taker_direction)?;
+    let amm_lp_allowed_to_jit_make = market
+        .amm
+        .amm_lp_allowed_to_jit_make(amm_wants_to_jit_make)?;
+    let split_with_lps = amm_lp_allowed_to_jit_make && amm_lp_wants_to_jit_make;
+
+    if amm_wants_to_jit_make {
+        liquidity_split = if split_with_lps {
+            AMMLiquiditySplit::Shared
+        } else {
+            AMMLiquiditySplit::ProtocolOwned
+        };
+
+        jit_base_asset_amount = calculate_jit_base_asset_amount(
+            market,
+            base_asset_amount,
+            maker_price,
+            valid_oracle_price,
+            taker_direction,
+            liquidity_split,
+        )?;
+    } else if split_with_lps {
+        liquidity_split = AMMLiquiditySplit::LPOwned;
+
+        jit_base_asset_amount = calculate_jit_base_asset_amount(
+            market,
+            base_asset_amount,
+            maker_price,
+            valid_oracle_price,
+            taker_direction,
+            liquidity_split,
+        )?;
+    }
+
+    Ok((jit_base_asset_amount, liquidity_split))
 }