+++ test-scripts/run-anchor-tests.sh
@@ -4,6 +4,7 @@ if [ "$1" != "--skip-build" ]; then
 fi
 
 test_files=(
+  perpLpJit.ts
   spotSwap.ts
   maxLeverageOrderParams.ts
   multipleMakerOrders.ts